+++ tests/ast/test_basic.py
@@ -10,6 +10,7 @@ from pyccel.ast.literals import LiteralInteger
 from pyccel.ast.operators import PyccelAdd, PyccelMinus, PyccelMul, PyccelOperator
 from pyccel.ast.variable import Variable
 from pyccel.errors.errors import Errors
+from pyccel.naming import name_clash_checkers
 from pyccel.parser.parser import Parser
 
 base_dir = os.path.dirname(os.path.realpath(__file__))
@@ -17,7 +18,11 @@ path_dir = os.path.join(base_dir, "scripts")
 
 
 def get_functions(filename):
-    pyccel = Parser(filename, output_folder=os.getcwd())
+    pyccel = Parser(
+        filename,
+        output_folder=os.getcwd(),
+        name_clash_checker=name_clash_checkers["python"],
+    )
     errors = Errors()
 
     ast = pyccel.parse(verbose=0)